# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Account API",
    "version": "1.0.0",
    "description": "API documentation for user accounts"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/account/create/{name}/{email}/{password}": {
      "post": {
        "summary": "Create a new user account",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "path",
            "description": "Password of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/account/login/{email}/{password}": {
      "get": {
        "summary": "Log in to user account",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "path",
            "description": "Password of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Login failed: user not found or wrong password"
          }
        }
      }
    },
    "/account/deposit/{email}/{amount}": {
      "post": {
        "summary": "Deposit amount to user account",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user to update balance",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "path",
            "description": "Amount to update balance. Positive value for deposit, negative value for withdrawal.",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User balance updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Insufficient balance for withdrawal."
          },
          "404": {
            "description": "User account not found"
          }
        }
      }
    },
    "/account/withdraw/{email}/{amount}": {
      "post": {
        "summary": "Withdraw amount from user account",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amount",
            "in": "path",
            "description": "Amount to withdraw from the user's account",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Withdrawal successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. User account not found or insufficient balance for withdrawal."
          }
        }
      }
    },
    "/account/find/{email}": {
      "get": {
        "summary": "Find user account by email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user to find",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User account found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User account not found"
          }
        }
      }
    },
    "/account/all": {
      "get": {
        "summary": "Get all user accounts",
        "responses": {
          "200": {
            "description": "User accounts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/account/delete/{email}": {
      "delete": {
        "summary": "Delete user account",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User account deleted successfully"
          },
          "404": {
            "description": "User account not found"
          }
        }
      }
    },
    "/account/logoff/{email}": {
      "get": {
        "summary": "Log off user",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email of the user to log off",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged off successfully"
          },
          "404": {
            "description": "User account not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          }
        }
      }
    }
  }
}